Add MIT License to the project
A new file, LICENSE, was added to the project which contains the MIT License. Adopting this license grants others permissions to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software subject to including the copyright notice. This helps clarify the terms under which others can use the project's code.