Update project and author details in Jekyll config file
The project's title, links (such as project URL and repository), and author details are added to the '_config.yml' file. This helps to provide further information about the author and the project while linking to the project's hosted webpage. It eliminated the initial redundancy in the author section. The relevant social media links were also provided for more author reachability.