Remove redundant whitespace and update repository URLs
Removed an extra line of white space from top of the 'gpt4-session-to-md.jq' file for cleaner code format. Updated the URL and repository field in the '_config.yml' file of the project docs.

These changes are to maintain consistency across all references to the project after the project repository's name was changed from 'gpt4-chat-session-to-markdown' to 'gpt4-session-to-markdown'. This ensures all links lead to the correct destination.